requirements.txt
# Enterprise Multimodal RAG - Airline Operations
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# LLM & Embeddings
openai>=1.12.0
tiktoken>=0.5.0

# Vector Store & Retrieval
faiss-cpu>=1.9.0
rank-bm25>=0.2.2

# Document Processing
pypdf>=4.0.0
pillow>=10.0.0

# Async & Streaming
httpx>=0.26.0
sse-starlette>=2.0.0
aiofiles>=23.0.0

# Monitoring & Logging
structlog>=24.1.0

# Data Processing
numpy>=1.24.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
python-dotenv>=1.0.0

